Add test of rulesDict1
